feat: delete message instead when only one attachment exists
